The WICED SDK's use of the 'ALTERNATE_MAIN' compile-time flag was buggy (evidently they never used it before)
ALTERNATE_MAIN is a way to pull the definition of "main( )" out of the
depths of the WICED SDK source tree, and put it someplace else. Like
perhaps in an app you're writing that wants to create some additional
FreeRTOS threads and the like.

If you say "-D ALTERNATE_MAIN" in the appropriate makefile -- in our case,
"WICED/RTOS/FreeRTOS/WICED/WICED.mk" -- you can de-activate the
standard definition of "main()" found in "WICED/RTOS/FreeRTOS/WICED/wiced_rtos.c".
That's the change to WICED.mk being commited here.

This should just work, but some declaration of things like FreeRTOS
thread handles that should've been declared in the scope of the code
controlled by the ALTERNATE_MAIN flag (because they naturally go
together with the "main()" definition), were not. So I moved some of
these declarations into that scope so that compiling an
externally-defined "main()" with the WICED SDK would work.